[PATCH] ppc32: handle misaligned string/multiple insns
This patch adds code to the ppc32 alignment exception handler to make it
handle the load/store string and load/store multiple word instructions.
This is an issue for older CPUs such as the PPC601, which traps on
load/store string instructions which cross a page boundary (newer CPUs
handle this in hardware).  I have a little test program which exercises
this code, so I am reasonably confident it's correct.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 412a4538XZ21O-DGB8kVLzGKpE9aHw