[PATCH] ppc64: Fix v_regs pointer setup
During some signal test, we found that v_regs pointer was not setup
correctly.  v_regs was made to point to itself, as a result of which the
pointer was corrupted when vec registers were copied over.  When the signal
handler returned, restore_sigcontext tried derefering the invalid pointer
and in the process killed the app with SIGSEGV.

Signed-off-by: Srivatsa Vaddagiri <vatsa@in.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 412a46716NLVd95kddGkHOn36vVWVA