[PATCH] ppc64: Use correct buffer size in RTAS call
Firmware expects the size of the buffer that you hand it when you ask it
for information about a hardware error to be of a very specific size, but
different versions of firmware appearently expect different sizes; using
the wrong size results in a painful, hard-to-debug crash in firmware.  Benh
provided a patch for this some months ago, but appreantly missed this code
path.  This patch sets up the log buffer size dynamically; it also fixes a
bug with the return code not being handled correctly.

Signed-off-by: Linas Vepstas <linas@austin.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 412a46ec51lSlKEB2SSPfcrvR4dfQA