[PATCH] ibmveth: module tag fixes
This and the following three patches contain bug fixes found in the
stabilization of SLES9.

This patch adds a call to MODULE_VERSION and changes the MODULE_AUTHOR call
to me (obviously with Dave Larson's permission).  It also increments the
version number to keep track of the bug fixes.  Please apply.

Signed-off-by: Santiago Leon <santil@us.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 412a470fqZhh18VvgUKplh7_EjB1UA