[PATCH] ibmveth: add memory barrier for hypervisor synchronisation
This patch adds a memory barrier to ensure synchronization with the
hypervisor (and avoid a panic when the hypervisor is halfway through
writing to the descriptor).  It also removes an unnecessary check that is
flawed anyway because the value can change between the atomic_inc() and the
assert.

Signed-off-by: Santiago Leon <santil@us.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 412a4732Qr56Fo4jdLY5tg1Dqg0Yrg