[PATCH] fbdev: ATTN: Maintainers - Set correct hardware capabilities
With David Eger's patch for advertising hardware capabilities to fbcon,
only a few drivers so far have been converted to do just that.  As a
result, scrolling speed of some drivers will not be optimal
(SCROLL_REDRAW).  The patch adds the correct flags for all drivers (except
for matroxfb -leave this to Petr-, and a few drivers that are not ported
yet to 2.6).

*Majority of changes is FBINFO_FLAG_DEFAULT -> FBINFO_DEFAULT

Signed-off-by: Antonino Daplas <adaplas@pol.net>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 412a4a81zINGQT62Qvu_KqxruLe4lA