[PATCH] sched: adjust p4 per-cpu gain
The smt-nice handling is a little too aggressive by not estimating the per cpu
gain as high enough for pentium4 hyperthread.  This patch changes the per
sibling cpu gain from 15% to 25%.  The true per cpu gain is entirely dependant
on the workload but overall the 2 species of Pentium4 that support
hyperthreading have about 20-30% gain.

P.S: Anton - For the power processors that are now using this SMT nice
infrastructure it would be worth setting this value separately at 40%.

Signed-off-by: Con Kolivas <kernel@kolivas.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 412b8458jS9nekxOzobFrmX7vMQB4Q