[PATCH] Make MAX_INIT_ARGS 32
We at Red Hat shipped a larger number of arguments for quite some time, it
was required for installations on IBM mainframe (s390), which doesn't have
a good way to pass arguments.

There are a number of reasonable situations that go past the current limits
of 8.  One that comes to mind is when you want to perform a manual vnc
install on a headless machine using anaconda.  This requires passing in a
number of parameters to get anaconda past the initial (no-gui) loader
screens.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 412b8788lpiBH6wjO3AmpeRqGV7RLg